requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.2
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-datatables-too==0.1.3
django-dotenv==1.4.2
django-extra-views==0.13.0
django-filter==2.4.0
django-multi-form-view==2.0.1
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.4.1
stripe==2.55.0
urllib3==1.25.11
